feat: Add weight field to Animal model
This commit adds the `Weight` field to the `Animal` model in both the `models.go` and `animal_handler.go` files. The `Weight` field is of type `float64` and is validated to be greater than or equal to 0. This change allows for the storage and retrieval of the weight of an animal in the database.